Fix tests for plaintext
